+++ mlflow/models/evaluation/artifacts.py
@@ -120,7 +120,6 @@ _InferredArtifactProperties = namedtuple(
     "_InferredArtifactProperties", ["from_path", "type", "ext"]
 )
 
-
 def _infer_artifact_type_and_ext(artifact_name, raw_artifact, custom_metric_tuple):
     """
     This function performs type and file extension inference on the provided artifact
